SIGAR API, Timer, First API caller and ConfigReader
DLL files outside of the SRC allow for the SIGAR API to work on Linux,
Mac OS and Windows x86 x64.  Basic API call has been implemented at the
moment. Everything is currently happening from the controller class to
give it a basic framework. A discussion needs to be had about OO for
this project as it is not really needed and this is quite linear. Basic
timer has also been added to repeat the task function after 10 seconds
for every 5 seconds.